# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "technique",
                "theory",
                "repertoire"
              ]
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "easy",
                "medium",
                "hard"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasksResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/routines": {
      "post": {
        "operationId": "RoutinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "get": {
        "operationId": "RoutinesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRoutinesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/api/v1/routines/{id}": {
      "get": {
        "operationId": "RoutinesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "patch": {
        "operationId": "RoutinesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "delete": {
        "operationId": "RoutinesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/api/v1/routines/{routineId}/tasks": {
      "post": {
        "operationId": "RoutinesController_addTask",
        "parameters": [
          {
            "name": "routineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRoutineTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineTaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "get": {
        "operationId": "RoutinesController_findTasks",
        "parameters": [
          {
            "name": "routineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineTaskWithTaskResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/api/v1/routines/{routineId}/tasks/reorder": {
      "patch": {
        "operationId": "RoutinesController_reorderTasks",
        "parameters": [
          {
            "name": "routineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRoutineTasksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineTaskResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/api/v1/routines/{routineId}/tasks/{taskId}": {
      "patch": {
        "operationId": "RoutinesController_updateTask",
        "parameters": [
          {
            "name": "routineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineTaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "delete": {
        "operationId": "RoutinesController_removeTask",
        "parameters": [
          {
            "name": "routineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
...
  },
  "info": {
    "title": "Guitar Coach API",
    "description": "API documentation for the Guitar Coach backend",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "category": {
            "type": "string",
            "enum": [
              "technique",
              "theory",
              "repertoire"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "referenceLink": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "category": {
            "nullable": true,
            "enum": [
              "technique",
              "theory",
              "repertoire"
            ],
            "type": "string"
          },
          "difficulty": {
            "nullable": true,
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "type": "string"
          },
          "referenceLink": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedTasksResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "category": {
            "enum": [
              "technique",
              "theory",
              "repertoire"
            ],
            "type": "string"
          },
          "difficulty": {
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "type": "string"
          },
          "referenceLink": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateRoutineDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "RoutineResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "taskCount": {
            "type": "number"
          },
          "totalTargetDurationMinutes": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "userId",
          "title",
          "taskCount",
          "totalTargetDurationMinutes",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedRoutinesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateRoutineDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "AddRoutineTaskDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number",
            "minimum": 1
          },
          "targetDurationMinutes": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "taskId"
        ]
      },
      "RoutineTaskResponseDto": {
        "type": "object",
        "properties": {
          "targetDurationMinutes": {
            "type": "number",
            "nullable": true
          },
          "routineId": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "routineId",
          "taskId",
          "position",
          "createdAt",
          "updatedAt"
        ]
      },
      "RoutineTaskWithTaskResponseDto": {
        "type": "object",
        "properties": {
          "targetDurationMinutes": {
            "type": "number",
            "nullable": true
          },
          "task": {
            "$ref": "#/components/schemas/TaskResponseDto"
          },
          "routineId": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "task",
          "routineId",
          "taskId",
          "position",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReorderRoutineTasksDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "taskIds"
        ]
      },
      "UpdateRoutineTaskDto": {
        "type": "object",
        "properties": {
          "position": {
            "type": "number",
            "minimum": 1
          },
          "targetDurationMinutes": {
            "type": "number",
            "minimum": 1
          }
        }
      },
...
    }
  }
}