# src/routines/routines.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  ParseUUIDPipe,
  Patch,
  Post,
  Query,
} from '@nestjs/common';
import { Session } from '@thallesp/nestjs-better-auth';
import type { UserSession } from '@thallesp/nestjs-better-auth';
import { AddRoutineTaskDto } from './dto/add-routine-task.dto';
import { CreateRoutineDto } from './dto/create-routine.dto';
import { FindRoutinesQueryDto } from './dto/find-routines-query.dto';
import { ReorderRoutineTasksDto } from './dto/reorder-routine-tasks.dto';
import {
  PaginatedRoutinesResponseDto,
  RoutineResponseDto,
} from './dto/routine-response.dto';
import {
  RoutineTaskResponseDto,
  RoutineTaskWithTaskResponseDto,
} from './dto/routine-task-response.dto';
import { UpdateRoutineDto } from './dto/update-routine.dto';
import { UpdateRoutineTaskDto } from './dto/update-routine-task.dto';
import { RoutinesService } from './routines.service';

@Controller('routines')
export class RoutinesController {
  constructor(private readonly routinesService: RoutinesService) {}

  @Post()
  create(
    @Session() session: UserSession,
    @Body() createRoutineDto: CreateRoutineDto,
  ): Promise<RoutineResponseDto> {
    return this.routinesService.create(session.user.id, createRoutineDto);
  }

  @Get()
  findAll(
    @Session() session: UserSession,
    @Query() query: FindRoutinesQueryDto,
  ): Promise<PaginatedRoutinesResponseDto> {
    return this.routinesService.findAll(session.user.id, query);
  }

  @Get(':id')
  findOne(
    @Session() session: UserSession,
    @Param('id', ParseUUIDPipe) id: string,
  ): Promise<RoutineResponseDto> {
    return this.routinesService.findById(session.user.id, id);
  }

  @Patch(':id')
  update(
    @Session() session: UserSession,
    @Param('id', ParseUUIDPipe) id: string,
    @Body() updateRoutineDto: UpdateRoutineDto,
  ): Promise<RoutineResponseDto> {
    return this.routinesService.update(session.user.id, id, updateRoutineDto);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  remove(
    @Session() session: UserSession,
    @Param('id', ParseUUIDPipe) id: string,
  ): Promise<void> {
    return this.routinesService.remove(session.user.id, id);
  }

  @Post(':routineId/tasks')
  addTask(
    @Session() session: UserSession,
    @Param('routineId', ParseUUIDPipe) routineId: string,
    @Body() addRoutineTaskDto: AddRoutineTaskDto,
  ): Promise<RoutineTaskResponseDto> {
    return this.routinesService.addTask(
      session.user.id,
      routineId,
      addRoutineTaskDto,
    );
  }

  @Get(':routineId/tasks')
  findTasks(
    @Session() session: UserSession,
    @Param('routineId', ParseUUIDPipe) routineId: string,
  ): Promise<RoutineTaskWithTaskResponseDto[]> {
    return this.routinesService.findTasks(session.user.id, routineId);
  }

  @Patch(':routineId/tasks/reorder')
  reorderTasks(
    @Session() session: UserSession,
    @Param('routineId', ParseUUIDPipe) routineId: string,
    @Body() reorderRoutineTasksDto: ReorderRoutineTasksDto,
  ): Promise<RoutineTaskResponseDto[]> {
    return this.routinesService.reorderTasks(
      session.user.id,
      routineId,
      reorderRoutineTasksDto,
    );
  }

  @Patch(':routineId/tasks/:taskId')
  updateTask(
    @Session() session: UserSession,
    @Param('routineId', ParseUUIDPipe) routineId: string,
    @Param('taskId', ParseUUIDPipe) taskId: string,
    @Body() updateRoutineTaskDto: UpdateRoutineTaskDto,
  ): Promise<RoutineTaskResponseDto> {
    return this.routinesService.updateTask(
      session.user.id,
      routineId,
      taskId,
      updateRoutineTaskDto,
    );
  }

  @Delete(':routineId/tasks/:taskId')
  @HttpCode(HttpStatus.NO_CONTENT)
  removeTask(
    @Session() session: UserSession,
    @Param('routineId', ParseUUIDPipe) routineId: string,
    @Param('taskId', ParseUUIDPipe) taskId: string,
  ): Promise<void> {
    return this.routinesService.removeTask(session.user.id, routineId, taskId);
  }
}